Add specific click API details, plus a code example
Now that we've solidified some of the API designs, we should have more details about them in the explainer where necessary. Getting from shared storage is pretty straightforward, but the click listener API needs an explicit description of its API surface. 

In addition, I've added an end-to-end example of how all of our API surfaces can be combined for a real-world use case of personalized payment buttons.