Add RTCPeerConnection constructor error to spec.
Because WebRTC's RTCPeerConnection interface is disallowed in
fenced frames, we need to patch the WebRTC spec to throw an exception
in the constructor. This patch also introduces the a "fenced
navigable ancestor" algorithm that can be used to determine if a given
document is nested in a fenced frame tree. This mechanism is used to
trigger the constructor error.