Add spec for APIs gated on network revocation.
After untrusted network access has been disabled in a fenced frame
via disableUntrustedNetwork(), script running in the frame will gain
access to new powerful APIs. This patch provides an algorithm that other
specs can use when implementing methods that should only succeed in a
fenced frame with network access revoked.

Currently, the only method that will rely on this algorithm is
SharedStorage.get(), which is implemented in the Shared Storage draft
spec. I'm planning to update that spec separately after this PR merges.